+++ README.md
@@ -1,2 +1,3 @@
-# syphilis_analyses
-Analyses of syphilis diagnostics
+# STIsim
+
+STI-specific modelling toolbox built on starsim. This repository contains disease modules, networks, and interventions for modelling a range of co-circulating STIs including HIV and syphilis.
